code/churn-model/db.py
import psycopg2
import psycopg2.extras
from contextlib import contextmanager
from .config import settings

@contextmanager
def get_conn():
    conn = psycopg2.connect(settings.pg_dsn)
    try:
        yield conn
    finally:
        conn.close()

def q(conn, sql, params=None):
    with conn.cursor(cursor_factory=psycopg2.extras.RealDictCursor) as cur:
        cur.execute(sql, params or ())
        if cur.description:
            return cur.fetchall()
        return None

def execmany(conn, sql, rows):
    with conn.cursor() as cur:
        psycopg2.extras.execute_batch(cur, sql, rows, page_size=1000)
