code/churn-model/requirements.txt
psycopg2-binary
pandas
scikit-learn
pydantic
python-dotenv
sqlalchemy
orjson
pydantic-settings
